+++ model/meal.go
@@ -6,10 +6,10 @@ type NewMeal struct {
 	ID          uuid.UUID `json:"-"`
 	Name        string    `json:"name"`
 	Jenis       string    `json:"jenis"`
-	Kalori      uint      `json:"kalori"`
-	Protein     uint      `json:"protein"`
-	Karbohidrat uint      `json:"karbohidrat"`
-	Lemak       uint      `json:"lemak"`
+	Kalori      float32   `json:"kalori"`
+	Protein     float32   `json:"protein"`
+	Karbohidrat float32   `json:"karbohidrat"`
+	Lemak       float32   `json:"lemak"`
 }
 
 type cariMeal struct {
